refactor(website): split homepage client component into server components
- Delete 828-line monolithic home-client.tsx client component
- Extract interactive islands: download-buttons.tsx, github-star-button.tsx
- Create server components: hero-section, feature-section, home-faq, company-section
- Convert news-section.tsx to server component (remove 'use client')
- page.tsx now composes everything server-side; enterprise + CTA inlined
- Fix DownloadButtons imports in 4 files that referenced deleted home-client
- All content now in initial server-rendered HTML for SEO
