+++ apps/website/src/app/(home)/_components/news-section.tsx
@@ -1,4 +1,3 @@
-'use client';
 import { ScrollReveal } from '@/components/landing/scroll-reveal';
 import { getNewsTypeBadgeLabel, type NewsType } from '@/lib/news';
 import { IconArrowRightFill18 } from 'nucleo-ui-fill-18';
